+++ tools/build-mingw64.sh
@@ -35,7 +35,10 @@ export PATH=$(pwd)/bin:${PATH}
 
 export OPAMROOT="$(cygpath -aml _opam)"
 
-opam init default "https://github.com/fdopen/opam-repository-mingw.git#opam2" -c "ocaml-variants.4.12.0+mingw64c" --disable-sandboxing --no-setup
+# If the following command fails with a curl error, make sure you have Cygwin's
+# curl in the PATH, and not a native Windows one.
+
+opam init default "https://github.com/fdopen/opam-repository-mingw.git#opam2" -c "ocaml-variants.4.14.0+mingw64c" --disable-sandboxing --no-setup
 
 eval $(opam env)
 
@@ -47,7 +50,7 @@ set -eu
 
 dune subst
 
-dune build -p ocamlformat
+dune build -p ocamlformat @install
 
 echo "Version check:"