+++ CHANGES.md
@@ -18,6 +18,7 @@
 - Fix alignment inconsistency between let-binding and let-open (#2187, @gpetiot)
 - Fix reporting of operational settings origin in presence of profiles (#2188, @EmileTrotignon)
 - Fix alignment inconsistency of if-then-else in apply (#2203, @gpetiot)
+- Fix automated Windows build (#2205, @nojb)
 
 ### Changes